use open factories to return SqliteConnections
