+++ lib/src/web/web_isolate_connection_factory.dart
@@ -1,12 +1,11 @@
 import 'dart:async';
 
 import 'package:sqlite_async/sqlite3_common.dart';
-import 'package:sqlite_async/src/common/abstract_open_factory.dart';
 import 'package:sqlite_async/src/common/isolate_connection_factory.dart';
 import 'package:sqlite_async/src/common/mutex.dart';
 import 'package:sqlite_async/src/common/port_channel.dart';
+import 'package:sqlite_async/src/sqlite_connection.dart';
 import 'package:sqlite_async/src/web/web_sqlite_open_factory.dart';
-import 'database/web_sqlite_connection_impl.dart';
 
 /// An implementation of [IsolateConnectionFactory] for Web
 /// This uses a web worker instead of an isolate
@@ -28,8 +27,8 @@ class IsolateConnectionFactoryImpl
   ///
   /// This opens a single connection in a background execution isolate.
   @override
-  WebSqliteConnectionImpl open({String? debugName, bool readOnly = false}) {
-    return WebSqliteConnectionImpl(mutex: mutex, openFactory: openFactory);
+  SqliteConnection open({String? debugName, bool readOnly = false}) {
+    throw UnimplementedError();
   }
 
   /// Opens a synchronous sqlite.Database directly in the current isolate.
@@ -43,8 +42,7 @@ class IsolateConnectionFactoryImpl
   ///     this connection.
   @override
   Future<CommonDatabase> openRawDatabase({bool readOnly = false}) async {
-    return openFactory
-        .open(SqliteOpenOptions(primaryConnection: false, readOnly: readOnly));
+    throw UnimplementedError();
   }
 
   @override
